+++ pkgs/google_generative_ai/CHANGELOG.md
@@ -1,4 +1,15 @@
-## 0.2.3-wip
+## 0.3.0
+
+- Allow specifying an API version in a `requestOptions` argument when
+  constructing a model.
+- Add support for referring to uploaded files in request contents.
+- Add support for passing tools with functions the model may call while
+  generating responses.
+- Add support for passing a system instruction when creating the model.
+- **Breaking** Added new subclasses `FilePart`, `FunctionCall`, and
+  `FunctionResponse` of the sealed class `Part`.
+
+## 0.2.3
 
 - Update the package version that is sent with the HTTP client name.
 - Throw more actionable error objects than `FormatException` for errors. Errors